// src/lib/types.ts
export type UserRole = 'admin' | 'manager' | 'waiter' | 'user';
// ...
export interface MenuItemIngredient {
  stockItemId: string;
  quantity: number; // The amount of the stock item used in this menu item
}

export interface MenuItem {
  id: string;
  name: string;
  price: number;
  category: string;
  imageUrl: string;
  isAvailable: boolean;
  ingredients: MenuItemIngredient[];
  costOfGoods: number;
}
// ...
export interface DecodedToken {
  id: string;
  name: string;
  role: UserRole;
  email: string;
  iat: number;
  exp: number;
}